[Change] separate image controllers from routes
[Fix] Remove end sql connection timeout
