+++ routes/imageRoutes.js
@@ -6,7 +6,8 @@ const { v4: uuidv4 } = require("uuid");
 const router = express.Router();
 
 const { catchRequestError } = require("../utils/catchAsync");
-const { dbPoolPromise, checkDB } = require("../mySqlConnection");
+const { dbPoolPromise } = require("../mySqlConnection");
+const image = require("../controllers/image_controller");
 
 const storage = multer.diskStorage({
     destination: (req, file, callBack) => {
@@ -24,38 +25,10 @@ const upload = multer({
     }
 });
 
-router.post("/addimages", upload.array("images", 5), (req, res) => {
-    const fileList = req.files;
-    res.send(fileList);
-    fileList.forEach(async (file) => {
-        const [saveImage] = await dbPoolPromise.execute(
-            "INSERT INTO images (id, name, source) VALUES (?, ?, ?)",
-            [file.filename, file.originalname, file.path]
-        );
-    });
-});
+router.post("/addimages", upload.array("images", 5), image.add);
 
-router.post(
-    "/deleteimages",
-    catchRequestError(async (req, res) => {
-        const imgUrl = req.body.url;
-        const newUrl = JSON.stringify(`./${imgUrl.replace(/\\/g, "/")}`);
-        fse.remove(imgUrl, (err) => {
-            if (err) return console.log(err);
-        });
-        const [delImg] = await dbPoolPromise.execute(
-            "DELETE FROM images WHERE source = ?",
-            [imgUrl]
-        );
-    })
-);
+router.post("/deleteimages", catchRequestError(image.delete));
 
-router.get(
-    "/getimages",
-    catchRequestError(async (req, res) => {
-        const [imageList] = await dbPoolPromise.execute("SELECT * FROM images");
-        res.send(imageList);
-    })
-);
+router.get("/getimages", catchRequestError(image.loadList));
 
 module.exports = router;
